+++ src/data/technicalindicator/directionalmovement/ExponentialMovingAverage.js
@@ -12,7 +12,7 @@
  * limitations under the License.
  */
 
-import TechnicalIndicator from '../TechnicalIndicator'
+import TechnicalIndicator, { TechnicalIndicatorSeries } from '../TechnicalIndicator'
 
 import { EMA } from '../defaultTechnicalIndicatorType'
 
@@ -20,7 +20,9 @@ export default class ExponentialMovingAverage extends TechnicalIndicator {
   constructor () {
     super({
       name: EMA,
+      series: TechnicalIndicatorSeries.PRICE,
       calcParams: [6, 12, 20],
+      precision: 2,
       shouldCheckParamCount: false,
       shouldOhlc: true,
       plots: [