+++ docs/en/CHANGELOG.md
@@ -8,6 +8,7 @@ New versions are generally released in about two weeks (emergency versions that 
 + 🆕 Add back the technical indicator ohlc display.
 + 🆕 Add text `weight` configuration.
 + 🆕 A new `shouldFormatBigNumber` property is added to the custom technical indicators, which is used to inform the chart whether larger numbers need to be formatted.
++ 🆕 New api `setTechnicalIndicatorPrecision`.
 + 💄 Optimize the data display after moving the mouse over the chart, and display the last data instead.
 + 💄 Optimize the starting position of chart drawing when loading new data.
 + 💄 Optimize the internal dom element level to reduce nesting.
@@ -15,9 +16,9 @@ New versions are generally released in about two weeks (emergency versions that 
 + 🐞 Fix the problem that illegal technical indicator calculation parameters may cause the chart to crash.
 + 👉 The style configuration `yAxis.tickText.position` is changed to `yAxis.inside`.
 + 👉 The default font family is changed from `Arial` to `Helvetica Neue`.
++ 👉 The custom technical indicator attribute `isPriceTechnicalIndicator` becomes `series`, and the value is changed to `price`, `volume` and `normal`.
 + 🗑 Remove the x-axis `minHeight` and `maxHeight` configuration, change to height adaptive.
 + 🗑 Remove the y-axis `minWidth` and `maxWidth` configuration, change to width adaptive.
-+ 🗑 Remove the custom technical indicator `isPriceTechnicalIndicator` attribute.
 
 ## 5.3.0
 `2020-05-26`